.github/utils/requirements.txt
python-slugify
pyyaml
